resolving paths and reading from them is now a job of the World, not the Manifest
