+++ Cargo.toml
@@ -10,6 +10,7 @@ path = "src/main.rs"
 [dependencies]
 async-trait = "0.1.80"
 clap = { version = "4.5.7", features = ["derive", "env"] }
+derive_more = { version = "2.0.1", features = ["debug"] }
 itertools = "0.14.0"
 once_cell = "1.19.0"
 reqwest = "0.12.5"
